DuckDB bindings
