+++ docs/api/c/arrays.rst
@@ -1,0 +1,20 @@
+Arrays
+======
+
+.. c:autotype:: vx_array
+    :file: vortex.h
+
+.. c:autofunction:: vx_array_clone
+    :file: vortex.h
+
+.. c:autofunction:: vx_array_free
+    :file: vortex.h
+
+Properties
+----------
+
+.. c:autofunction:: vx_array_len
+    :file: vortex.h
+
+.. c:autofunction:: vx_array_dtype
+    :file: vortex.h
